added guest test and practice option, fixed adding words

+++ flashcards/viewmodel/editListViewModel.cs
@@ -5,6 +5,7 @@ using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows;
 
@@ -20,12 +21,19 @@ namespace Flashcards.viewmodel
         public void getSelectedFile(string v)
         {
             List<string> fileContent = _FM.readFromFile(v, _profile);
-            foreach (Window window in Application.Current.Windows)
+            if (_profile != null && _profile.Length > 0)
             {
-                if (window.GetType() == typeof(MainWindow))
+                foreach (Window window in Application.Current.Windows)
                 {
-                    (window as MainWindow).DataContext=new editSelectedListViewModel(v,_profile);
+                    if (window.GetType() == typeof(MainWindow))
+                    {
+                        (window as MainWindow).DataContext=new editSelectedListViewModel(v,_profile);
+                    }
                 }
+            }
+            else
+            {
+                MessageBox.Show("You need to be logged in to edit files.");
             }
         }
 
@@ -37,14 +45,35 @@ namespace Flashcards.viewmodel
 
         public void addNewList(string listName)
         {
-            _FM.saveTextFile(listName, _profile);
-            refreshList();
+            if (_profile != null)
+            {
+                if (Regex.IsMatch(listName, "^[a-zA-Z0-9_]+$"))
+                {
+                    _FM.saveTextFile(listName, _profile);
+                    refreshList();
+                }
+                else
+                {
+                    MessageBox.Show("Invalid file name, it should contain only alphanymerical characters and underscore.");
+                }
+            }
+            else
+            {
+                MessageBox.Show("You need to be logged in to edit files.");
+            }
         }
 
         public void removeList(string fileName)
         {
-            _FM.deleteFile(_profile, fileName);
-            refreshList();
+            if (_profile != null && _profile.Length > 0)
+            {
+                _FM.deleteFile(_profile, fileName);
+                refreshList();
+            }
+            else
+            {
+                MessageBox.Show("You need to be logged in to edit files.");
+            }
         }
     }
 }
